[임지훈] 보관함 기능 Member 와 분리
